#12 Center the maze

+++ MazeWpfApp/Views/CellView.xaml.cs
@@ -21,7 +21,7 @@ namespace MazeWpfApp.Views
     /// </summary>
     public partial class CellView : UserControl
     {
-        public CellView(int topLeftX, int topLeftY, int sizeOfCell)
+        public CellView(double topLeftX, double topLeftY, int sizeOfCell)
         {
             InitializeComponent();
             DataContext = new CellViewModel(topLeftX, topLeftY, sizeOfCell);
